Adding Bremove:kqe_slot,kaipo_slot
One major sticking point for some FE players (hi Fleury) is that the
Hook route features a made-up boss at the KQ Eblan spot (which is
normally a cutscene fight like Kain-vs.-Cecil in Fabul). Another
sticking point for some FE players when it comes to boss hunts (hi
Inven) is that the Package boss (in Kaipo Inn, the Officer/Soldiers
fight) is very time-consuming and out of the way, and makes boss hunts
unfun/uninteresting. So, why not just remove those fights entirely?

Bremove does exactly that. It has to remove the boss slots for the
main assignment in core_rando.py, for auxiliary boss handling in
boss_rando.py, and for FuSoYa's location-based spell learning (only
on Scythe's fork+). It actually has to avoid removing those slots for
the Payable Golbez wacky, which is why that wacky now has a setup
phase. The FuSoYa thing also needs a bit of manual tweaking, in the
same vein. We now also have to actively count the boss spots, too, so
that's why randomizer.f4c now includes a number that will be substituted
for and why ending.f4c no longer hardcodes 34 bosses. The Hook route and
Kaipo Inn events also need to change, to bypass everything to do with
those bosses.

(It is very interesting how Python handles dictionaries/etc. being
changed and called in/by different modules.)