Expanding -exp:objectivebonus and -exp:geometric
The first attempt at bonus exp for objective completion and scaled exp
for repeated monster kills only had fixed parameters. This second
attempt now allows some choice of parameters for both flags. It also
cleans up a bit of code (thanks to Galeswift) and implements
Util__Divide to divide by other numbers instead of just powers of 2.

For the objective bonus, the suffix is used to substitute the correct
divisor into experience_acceleration.f4c (conveniently we already
have a substitution entry for the number of objectives in the seed if
necessary).

For the geometric modifier, the utility function is moved to
randomizer_boss.f4c to go with the other geometric exp modifications.
The suffix to the flag is used to substitute the correct multiplier
for the utility function.

The UI and flag descriptions have been updated, as has flagsetcore.
The f4c substitutions for the code are implemented with the "toggle"
functionality.