Adding -tweak:darkpaladin
Many folks have asked "What if Cecil just became a stronger Dark
Knight after Ordeals?" In the romhack Unprecendented Crisis, you have
the option to stay as DKC the whole game, and the weaponry/stats for
DKC are set up for endgame DKC to be perfectly viable. This tweak does
something similar, within the confines of the FE/f4c framework.

Generally speaking, this flag turns Paladin Cecil from a beefy tank
with a good Str stat and the best weapons in the whole game, convenient
Agi for Zeromus, and utility white magic into a much-less-beefy
offensive juggernaut with inconveniently high Agi and a selection of
utility/status black magic as well as Nuke (since it's single-target),
together with enough Wisdom to sort of contribute effectively (but not
really enough MP to do it all himself)... and still get the good gear.
In lieu of AoE black magic (e.g. Quake), he gets the Dark Wave
command, which is unchanged. That said, Dark Paladin Cecil gets
all of his usual equips, so his attack stat can be much higher, and so
Dark Wave is actually a viable option now. The graphics turn out to
just... work, which is appreciated. In addition, Cecil will still Cover
units in critical HP, though he cannot manually Cover.

All of the above changes were fairly straightforward to make, via
direct patching (with loops for e.g. the level-up table) and the
spellset f4c compiler. What was *not* straightforward were the many,
many equipment changes required to maintain the theme, because FE uses
prebuilt item description text and pre-assembled item databases. So,
this commit features no fewer than four instances where spoiler entries
are loaded and then if this flag is on, they get changed before they
are entered into the spoiler table. Changing the databases themselves
would have required building/calling new copies of the database
functions, which sounded worse than what I ended up doing. As it
stands, there are new database/txt files that generate.py used to
build the binaries for the item descriptions, and then generator.py
just loads the correct pre-built binaries.

Lastly, Lightbringer has been changed to Deathbringer; it casts Fatal
and (because why not) it now hits dragons for weakness instead of
spirits and undeads. None of the graphics for any equipment have been
changed, so it'll look weird, but whatever. No dialog has been changed
in-game either (so it will still say that Cecil became a Paladin), but
the job name is now "DPaladn".