Updating spoiler log entries for -tweak:kainmagic
If FuSoYa learns spells with this flag active, then it would still
say "Sight" in the spoiler log instead of "Lance". This behaviour
has now been fixed (unfortunately requiring the same two lines of
code in multiple places, but at least list comprehensions are cool;
there might be a better way to do it though).