"Correcting" encounter level calculations/etc
Both vanilla FF4 and FE have quirks/bugs with calculating average
levels, in different places. These changes are meant to align all
averaging calculations so that the behaviour is the same regardless
of flag context and are in line with what, presumably, was the
original FF4 developer intention.

First, vanilla FF4 manages to only check the first three monster slots
when computing the average monster level of a formation. The one-byte
patch in encounter_average_level.f4c fixes that issue (thanks Aexoden!).
FE uses the vanilla code when on Evanilla, so the above patch is
actually useful. This patch is loaded in encounter_rando.py.

Second, FE moves the average level check to before the battle starts
when Edanger is on. The way it checks is to preload the encounter,
load precalculated formation levels, and then run the checks before
deciding to do the battle or not. The precalculated formation levels
are computed *without* any monsters that are called (or transformed
into, but that only happens in boss battles) but still including all
eight monster slots. On the other hand, the game's vanilla code does
not skip pre-swooned/called monsters (when they are in the first three
slots, anyway). In generate_formation_levels_mod.py, the precalculated
formation levels are computed including all non-empty monster slots,
the way the vanilla game would have... if it didn't just check the
first three slots, which is now what happens in Evanilla anyway. The
new levels are in formation_average_levels_mod.bin, which is loaded
by generator.py. (generator.py is not included in this commit, because
it has other changes for different flags in an upcoming commit.)

Third, in encounter_dangerous.f4c there is a line of code that, when
computing your average party level, divides by 5 instead of by the
number of characters you have, thus artificially lowering your
average party level. That code has been changed to actually compute
the average correctly.