Adding more -exp: flags
This update includes a number of new experience flags (inspiration
from the fork-feedback-and-ideas forum experience topic). It also
revamps the experience_acceleration.f4c file and adds in the relevant
code to core_rando.py and generator.py.

The main structural changes to the f4c file are that all of the
overflow checking happens via jumping ahead to the code that sets your
experience to 0xFFFFFF and the addition of a general-purpose function
that takes in a divisor in the accumulator and an index in the Y
register and adds to your current experience a fraction of that
experience Y times, thus allowing a variety of "add X % exp per
[criteria]" flags. "X" right now can only be "100 // integer" values.

-exp:crystalbonus checks if you have the Crystal and then awards
double exp, just like the 10 KI bonus.

-exp:maxlevelbonus looks at the difference between 2x your max party
level and the minimum monster level; for every gap of 5 levels, award
20% bonus experience.

-exp:kicheckbonus reimplements from scratch the KI check bonus flag
that Galeswift made, no longer piggybacking off -exp:objectivebonus.
rewards.f4c now always defines Rewards__KeyItemChecksCount at $7e162b
(the same place Galeswift eventually chose but with a different name),
and it counts the number of *potential* key item checks completed
(code in rewards.f4c changed for both quest checks and MIAB checks).
In particular, it does not count checks which are not included in
the key item pool; if Kmoon is not on, for instance, then checking
the CS altar will not increase this counter (which is distinct from
the original flag). The check in rewards.f4c is based on the data
provided for the purposes of Rosa's Mom hints, meaning that when
flags are hidden, every K spot is a potential KI check, including all
MIABs. This flag does still ignore the starting key item check.

-exp:zonkbonus awards an exp bonus for zonking out of a KI check.
The flag relies on the same Rewards__KeyItemChecksCount used by
-exp:kicheckbonus, which is why rewards.f4c now always defines it.
It takes that number and subtracts off the number of key items already
obtained; the result *should* be the number of zonks, but since we
are not counting the starting check (which could be a KI), we need
to possibly subtract 1 from the number of key items... which amounts
to setting the carry before the subtraction (or not). There was no
real point to making an equivalent _num check, since "getting all
zonks for max exp" feels uninteresting.

-exp:miabbonus awards bonus experience for MIAB encounters, checking
the same way that Enoexp checks (seeing if the chest item is non-zero).

-exp:moonbonus awards bonus experience for moon encounters; the check
works by checking map plane for the moon's surface and then checking
for indoor maps with map plane underworld/moon and map ID starting
with Cave Bahamut 1F; all moon maps are beyond that point.

Changes were made to ensure that the generator page doesn't break (by
removing all capital letters from flags like -exp:objectivebonusN...
whoops).

core_rando.py was modified to fix a v4.6.0 bug where the LST MIABs
were not removed from potential key item consideration under Kmiab
without Kmoon/unsafe. The number of potential key item slots was
added to env.meta for -exp:kicheckbonus_num.

generator.py was modified to include the new formation changes from
the previous commit, as well as build the substitutions for the
exp flags.