Adding Omode:external
FE obviously does not have in-game support for things like bingo or
meta objectives or similar. However, we can repurpose a Very Useful
Item (the EagleEye; it's not actually useful) to instead check off
an objective when you're completed your external goals. The item
should be usable when you're not on the overworld, due to how
textboxes work. Theoretically.

The long story short is that the EagleEye in FE is actually the "Smut"
(the dirty magazine) that's in the J version of FF4, but normally it
just points to the item event that casts Sight. Instead, we let it
call its normal event ($DB), but instead change that event to mark
an objective as complete.

To accomplish all of this, we add an objective mode (so we need to
modify objectivespec and flagspec), add capacity for a fifth starter
kit (up from four; normally you only get three, but the Save Us Big
Chocobo wacky bumps it to four, so we need a fifth here) by using
another unused message ($A6), change the opening to give us the kit,
modify a whole bunch of sellable/trashable checks (discardable relies
on sellable), remove the EagleEye from shops and treasures, and
finally write the event that completes the objective. The call_orbs.f4c
script usually changes "Smut" into EagleEye; we tell it not to do that.

It's not obvious in which maps the original item could be used, but it
turns out that you could use the "Smut" in every non-overworld map,
which is exactly what we want from our EagleEye. So? Do nothing.