Removing Cursed Rings from Cthrift
Here are some reasons why preventing characters from joining your
party already wearing a Cursed Ring is a good idea:
- It prevents a 1/8 potential softlock for Hero Cecil fights in the
changing room on Ordeals on Cthrift:4. Having a significant agility
disadvantage and all-1s stats makes that fight effectively unwinnable
for a non-trivial number of bosses at that spot.
- Cursed Rings are the most valuable tier 4 item, by 21000 GP as of
v4.6.0. They are also the single most valuable anchoring item in the
game. Getting one for free is game-warping (and not in a good way),
and you often get more than one because of dupes (if enabled) and the
relative lack of arm equipment for most characters.
- It preserves the flagset designer's choice when considering to use
the -nocursed flag.

While it's funny to have characters start with a Cursed Ring and it
would be a Bad-Ideas-type situation to have to manage the softlock
fight, overall the benefits of removing Cursed Rings outweigh the cons.

The implementation here is simply to refine the arms dbview to
exclude Cursed Rings. The uispec is updated to note this exclusion.