Adding -agility flags and -speedmodbalance
This update is a big one, expanding the options for agility from
"vanilla, hero, or normal FE" to a whole bunch of things. The changes
take place in the agility.f4c file, which took the anchoring mod out
of character_expansion.f4c and greatly expanded it. generator.py also
has some on-the-fly substitutions and toggles.

The new flags are:

-agility:vanilla, which is just a renaming of the original vanilla
agility flag. Two unnecessary lines of code were removed because they
duplicated an empty slot check.

-agility:slowest and -agility:fastest, which find the slowest and
fastest characters in your party, respectively, and use them as the
anchors.

-agility:median, which does a calculation to find the slot with the
median agility in your party.

-agility:average, which computes the average agility of your party
and uses that number for anchoring purposes.

-agility:monster, which does the same thing but for the monster
formation. This flag is mostly a joke/super hard challenge flag.

-agility:flat, which treats every character/monster as having 1 Agi,
so every actor gets the same RA.

-agility:750formula, which implements an RA formula created by S3.

-agility:anchor[7,27,28,41,42], which uses those specific agility
values for the anchoring calculation. The numbers were chosen for the
following reasons: 7, to let even Fu be RA1 at endgame; 27 and 41, to
make Zeromus RA1 or RA2 with the worst anchoring for your party; and
28 and 42, to make Zeromus RA2 or RA3 with the best anchoring for your
party.

-agility:scale[1,10], which scales the anchor timer from 5 to 1 or 10.
This flag can combine with any of the above (or with Chero).

Some of the flags modify the Count timer (used by Plague and FatalEye
and sometimes Wyvern under whichburn), because the base timer is too
short to let you get enough actions off to win otherwise.

This update also adds -speedmodbalance, which is also an S3 creation.
It changes the lower bound of the speed modifier from 12 to 8, makes
Fast and Slow change the speed modifier by 4 instead of 3 and 8, and
adjusts SilkWebs to be double Slow (Hermes is already double Fast!).
This isn't quite S3's vision, but it's more fun this way I think. The
changes here are in speed_modifier.f4c.

flagspec, uispec, and uispec_flagdescriptions have been updated.