Fixing -exp:zonkbonus[n]
There were two issues with the -exp:zonkbonus[n] flags. (Reminder: a
"zonk" is a non-key-item reward from a potential key item check, with
the exception of the starting item slot because there's no choice
about making that check.)
1. MIABs did not correctly count key item checks even if they were
potential key item locations.
2. The computation did not take into account the fact that you can
acquire the Crystal without completing a check, which would decrease
the number of zonks (or underflow it, on some flags). The same thing
happens on other forks, e.g. via gated objectives on Galeswift's fork.

The first issue is fixed by making the key item check counting a
permanent piece of the code in rewards.f4c. The second issue... needs
a bit more effort. In the end, I made a "zonk counter" to match the
"key item checks" counter (on this fork, at $7e162c) and wrote a bit
more code to correctly increment that counter when delivering zonk
rewards from key item slots.

For quest rewards, we simply move the relevant counter checks to
after the key-item-specific code. If we execute that code, then we
increment the key item checks counter because you can't get a key item
from a non-key-item check, of course. If not, then either the quest
wasn't a potential key item check at all, or it was and it's also a
zonk.

For MIABs, it's a bit more of a mess. We leave the potential key item
check at the beginning, but push the check value (0x01 or 0x00) onto
the stack for later. I'm not sure what "ExternalRewards" are, but
I'm willing to assume that they're not zonks, so in that case and in
the case that we get a key item, then we pull the check value off the
stack and carry on as usual. Otherwise, in the "regular item" code we
pull the check value off the stack and use it to check if we need to
also increment the zonk counter.

To make use of the counter, experience_acceleration.f4c was changed
(note that we still need to subtract 1 from the zonk counter, because
it does not take into account our rule that a starting non-key-item
isn't actually a zonk). generator.py was tweaked a little bit to make
the code more readable.