Bugfix for Wisdom/Will-based timers
This bugfix, contained in fix_wisdom_will_timers.f4c, addresses
two vanilla bugs:

1. The sap timer calculation does a multiplication. In the process,
it stores a value into the hi byte of a two-byte parameter instead of
the lo byte, and the function doesn't use the hi byte at all (it is
immediately set to 0 in a multiplication routine). So, the first patch
is to store that value into the lo byte instead. Now the timer will
correctly be set to 2*(Wis+Wil) + 30 or 4*($2F stat) + 30, instead of
30 + (2 or 4)*a random value.

2. The paralyze/sleep/gradual petrification (for... the Medusa Sword,
okay, not super relevant) timer has an arithmetic bug where it
multiplies a byte by 4 (Will or $2F stat) but does not take care of
the top bit of the result. Here, we have written some new code to
properly take care of the multiplication and resulting subtraction, so
that the timer's base value is actually 300 - 4*stat. (After that,
it is clamped to be positive, adjusted by the speed modifier, and
then is divided by 6 to get the final result.)

generator.py is updated to include the f4c file.