Adding -wacky:scrambledstats
Idea partially due to Coffee&Chocobos: What if Agi and Str played
opposite rolls in the calculation of attack multipliers? Meaning,
what if it were Agi/8 + Str/16 + 1? This wacky goes multiple steps
further, shuffling the roles of all five basic stats (Str, Agi, Vit,
Wis, and Wil).

The implementation is, in all but one case, to find every load of a
stat and replace the lo byte of that load (they're all two-byte loads)
with the byte for the stat to be used for that calculation. The
shuffling is consistent; for example, if Str -> Agi, then everywhere
Agi is normally used, Str will be used instead. The one other case
is when magic damage is being calculated; there, the vanilla code
sets an offset to go "+1" if the spell is white magic to access Wil
instead of Wis, and in our code we set it to be the appropriate offset
from Str (since the offset is 0-4 and can't be negative, as it's
going into the X register).

agility.f4c gets a non-trivial revision, to handle what happens when
the stat to be used for calculating agility isn't agility anymore. The
issue, of course, is that monsters are exempt from stats-shuffling, so
when a monster is being checked, their agility needs to be normal.
So, we define CharAgiAddress to be the appropriate stat address via
substitution with default value 2015, and some new code is (almost)
always added to split up monsters vs. characters when checking slot
agility. The "750 formula" agility code is slightly rewritten to
accommodate that change as well.

Thanks to Wylem for writing the character-stats-explained wiki page,
which outlined every stat usage and made it less scary. (It should go
without saying that Aexoden is the GOAT for the disassembly.)