Adding Enomacgiant
This flag is practically identical to sgrunt's "Enodmachin" flag,
which removes the repeatable D$ grind encounter by replacing the
D.Machin with a Horseman. This flag replaces the MacGiant in the
repeatable MacG grind encounter with a Machine (and fixes the
cursor data, etc.). The relevant f4c file is added in
encounter_rando.py.